Brand 도메인 수정 1. 멤버필드 wish count 제거(wish_brand와 데이터 중복) 2. 상위 10 브랜드 response 수정(수정 전:브랜드 이름 중복 출력됨) 3. delete 요청 url 수정 4. 브랜드 조회수 증가 엔드포인트 구현

+++ src/main/java/com/ward/ward_server/api/item/entity/Item.java
@@ -37,7 +37,7 @@ public class Item extends BaseTimeEntity {
     @JoinColumn(name = "brand_id")
     private Brand brand;
 
-    private int viewCount = 0;
+    private Long viewCount = 0L;
 
     @Column(nullable = false)
     @Convert(converter = CategoryConverter.class)
